requirements.txt
Flask==2.3.3
google-api-python-client==2.108.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-google-genai==0.0.5
openai>=1.6.1,<2.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
